Backend/requirements.txt
Django==5.2.5
djangorestframework==3.15.0
djoser==2.3.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
python-decouple==3.8
setuptools>=65.0.0 
python-docx
pandas==2.3.2
matplotlib==3.10.6
numpy==2.3.3
